pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bbx-workflow"
version = "1.0.0"
description = "Production-grade workflow automation platform with MCP integration"
readme = "README.md"
license = {text = "BSL-1.1"}
authors = [
    {name = "Ilya Makarov"}
]
requires-python = ">=3.10"
keywords = ["workflow", "automation", "orchestration", "devops", "cicd", "yaml", "dag", "mcp", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "click>=8.1.7",
    "pyyaml>=6.0.1",
    "httpx>=0.25.2",
    "pydantic>=2.5.2",
    "jinja2>=3.1.2",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
]
docker = ["docker>=6.1.3"]
aws = ["boto3>=1.34.0"]
browser = ["playwright>=1.40.0"]
ai = [
    "llama-cpp-python>=0.2.90",
    "tqdm>=4.66.1",
]
mcp = ["mcp>=1.0.0"]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.12.1",
    "ruff>=0.1.6",
    "mypy>=1.7.1",
]
full = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "docker>=6.1.3",
    "boto3>=1.34.0",
    "llama-cpp-python>=0.2.90",
    "tqdm>=4.66.1",
    "mcp>=1.0.0",
    "prometheus-client>=0.19.0",
]

[project.scripts]
bbx = "cli:cli"

[project.urls]
Homepage = "https://github.com/kurokie1337/bbx"
Documentation = "https://github.com/kurokie1337/bbx/tree/main/docs"
Repository = "https://github.com/kurokie1337/bbx"
Issues = "https://github.com/kurokie1337/bbx/issues"

[tool.setuptools.packages.find]
exclude = ["tests*", "examples*", "docs*"]

[tool.setuptools.package-data]
blackbox = ["py.typed"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
    | __pycache__
)/
'''

[tool.ruff]
line-length = 88
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
